requirements.txt
chromadb==0.5.3
numpy<2.0.0
sentence-transformers==2.7.0
google-generativeai==0.8.3
pandas==2.2.3
rouge-score==0.1.2
python-dotenv==1.0.1
